+++ src/main/java/com/smartx/tower/model/BackupPackage.java
@@ -62,7 +62,7 @@ public class BackupPackage {
 
   public static final String SERIALIZED_NAME_SIZE = "size";
   @SerializedName(SERIALIZED_NAME_SIZE)
-  private Double size;
+  private Long size;
 
   public static final String SERIALIZED_NAME_VERSION = "version";
   @SerializedName(SERIALIZED_NAME_VERSION)
@@ -232,7 +232,7 @@ public class BackupPackage {
   }
 
 
-  public BackupPackage size(Double size) {
+  public BackupPackage size(Long size) {
     
     this.size = size;
     return this;
@@ -245,12 +245,12 @@ public class BackupPackage {
   @javax.annotation.Nonnull
   @ApiModelProperty(required = true, value = "")
 
-  public Double getSize() {
+  public Long getSize() {
     return size;
   }
 
 
-  public void setSize(Double size) {
+  public void setSize(Long size) {
     this.size = size;
   }
